Changed - refactored workspace views (again)
