+++ src/ploneintranet/workspace/browser/configure.zcml
@@ -13,7 +13,7 @@
   <browser:page
       name="view"
       for="ploneintranet.workspace.workspacefolder.IWorkspaceFolder"
-      class=".views.WorkspaceView"
+      class=".workspace.WorkspaceView"
       permission="zope2.View"
       template="templates/workspace.pt"
       />
@@ -116,7 +116,7 @@
       name="ploneintranet_workspace_state"
       for="*"
       permission="zope.Public"
-      class=".views.WorkspaceState"
+      class=".workspace.WorkspaceState"
       allowed_interface="..interfaces.IWorkspaceState"
       />
